refactoring of event handler - new handler for app_home_opened event
