+++ src/handlers/events/channel/ChannelEvents.handler.ts
@@ -2,15 +2,13 @@ import { Response, Request } from 'express';
 import Logger from '../../../utils/Logger';
 import ResponseHandler from '../../ResponseHandler';
 import Channel from '../../../models/Channel/Channel.model';
+import EventHandler from '../Event.handler';
 
-export default class ChannelEventHandler {
-    rh: ResponseHandler;
-    constructor() {
-        this.rh = new ResponseHandler();
-    }
+export default class ChannelEventHandler extends EventHandler {
+    constructor() { super(); }
 
     /** For when the bot is kicked out of a channel */
-    public async handleChannelLeft(req: Request, res: Response): Promise<void> {
+    public async handleEvent(req: Request, res: Response): Promise<void> {
         let channel_removed = req.body.event.channel;  // ID of the channel this bot was removed from
         let channel;
