+++ src/utils/Router.ts
@@ -1,10 +1,8 @@
-import EventHandler from "../handlers/events/app_mentions/Events.handler";
+import EventHandler from "../handlers/events/app_mentions/AppMentionHandler.handler";
 import ChannelHandler from "../handlers/events/channel/ChannelEvents.handler";
 import logger from './Logger';
 import { Response, Request } from 'express';
-
-const ev = new EventHandler();
-const ch = new ChannelHandler();
+import AppHomeOpenedHandler from "../handlers/events/app_home_opened/AppHomeOpened.handler";
 
 /**
  * Given a request, route it to the appropriate handler.
@@ -15,11 +13,15 @@ const ch = new ChannelHandler();
 export default async function RouteRequest(requestType: string, req: Request, res: Response) : Promise<void>{
     switch (requestType) {
         case "app_mention":
-            ev.appMention(req, res);
+            new EventHandler().handleEvent(req, res);
             break;
         
         case "channel_left":
-            ch.handleChannelLeft(req, res);
+            new ChannelHandler().handleEvent(req, res);
+            break;
+        
+        case "app_home_opened":
+            new AppHomeOpenedHandler().handleEvent(req, res);
             break;
     
         default:
